[#50] SwiftDoc 정리, VM 파일 분리
