+++ PomoTodo/PomoTodo/Sources/Views/MainTabBarView.swift
@@ -55,11 +55,6 @@ struct MainTabBarView: View {
       setupTabBarAppearance()
       _ = container.pomoTodoUseCase.getTodayPomoDay()
     }
-//    .onChange(of: selectedTab) { oldTab, newTab in
-//      if newTab != .Pomo {
-//        // pomoVM.saveTomatoProgress()
-//      }
-//    }
   }
   
   //MARK: - Funcs
