+++ PomoTodo/PomoTodo/Sources/Views/Pomo/Subviews/ActiveSubviews/CircularProgressBar.swift
@@ -7,17 +7,20 @@
 
 import SwiftUI
 
+/// 원형 진행 바 (타이머 진행률 표시)
+/// - 배경 원 (고정된 회색 원)
+/// - 진행률 원 (타이머 진행 상태 표시)
 struct CircularProgressBar: View {
   @EnvironmentObject var pomoVM: PomoViewModel
   
   var body: some View {
     ZStack {
-      // background Circle
+      /// 배경 원 (고정된 회색 원)
       Circle()
         .stroke(lineWidth: 10)
         .foregroundStyle(.timerGray)
       
-      // progress Circle
+      /// 진행률 원 (타이머 진행 상태 표시)
       Circle()
         .trim(from: 0.0, to: pomoVM.progress)
         .stroke(pomoVM.selectedColorSet.normalColor, style: StrokeStyle(lineWidth: 10, lineCap: .round))
@@ -27,7 +30,3 @@ struct CircularProgressBar: View {
     .frame(width: 280, height: 280)
   }
 }
-
-//#Preview {
-//    CircularProgressBar()
-//}
