+++ PomoTodo/PomoTodo/Sources/Views/Pomo/Subviews/ActiveSubviews/TimerDisplayView.swift
@@ -7,12 +7,16 @@
 
 import SwiftUI
 
+/// 타이머 정보 뷰 (남은 시간 및 토마토 진행 상태)
+/// - 디지털 시계 스타일로 남은 시간 표시
+/// - 현재 사이클 진행 상태를 아이콘으로 표시
 struct TimerDisplayView: View {
   @EnvironmentObject var pomoVM: PomoViewModel
   
   var body: some View {
     VStack(alignment: .center) {
-      Text(timeString(from: pomoVM.remainingTime))
+      /// 남은 시간 표시
+      Text(pomoVM.remainingTime.asTimeString)
         .monospacedDigit()
         .font(.pretendard(.extraBold, size: 56))
         .frame(width: 280, alignment: .center)
@@ -20,6 +24,7 @@ struct TimerDisplayView: View {
       
       Spacer().frame(height: 32)
       
+      /// 토마토 진행 상태 아이콘
       HStack(alignment: .center, spacing: 8) {
         tomatoIcons()
       }
@@ -28,14 +33,9 @@ struct TimerDisplayView: View {
     } // : vstack
   }
   
-  //MARK: - UI settings funcs
-  
-  private func timeString(from time: Int) -> String {
-    let minutes = time / 60
-    let seconds = time % 60
-    return String(format: "%02d : %02d", minutes, seconds)
-  }
-  
+  // MARK: - UI Helper Methods
+    
+  /// 토마토 진행 상태 아이콘 표시
   @ViewBuilder
   private func tomatoIcons() -> some View {
     ForEach(0..<(pomoVM.curTomato - 1), id: \.self) { _ in
@@ -51,6 +51,7 @@ struct TimerDisplayView: View {
     }
   }
   
+  /// 토마토 아이콘 생성 (색상 및 크기 통일)
   @ViewBuilder
   private func createTomatoIcon(systemName: String) -> some View {
     Image(systemName: systemName)
@@ -60,7 +61,3 @@ struct TimerDisplayView: View {
       .frame(height: 15)
   }
 }
-
-//#Preview {
-//    TimerDisplayView()
-//}
