+++ PomoTodo/PomoTodo/Sources/Views/Pomo/Subviews/ActiveSubviews/TimerRunningInnerView.swift
@@ -7,12 +7,14 @@
 
 import SwiftUI
 
+/// **현재 선택된 태그 라벨 뷰**
+/// - 포커스 상태일 때는 태그명만 표시
+/// - 휴식 시간일 때는 아이콘과 함께 표시
 struct TimerTagLabelView: View {
   @EnvironmentObject var pomoVM: PomoViewModel
   
   var body: some View {
     HStack(alignment: .bottom) {
-      // 휴식 시간일 때만, 아이콘 표시
       if pomoVM.currentPhase != .focus {
         Image(systemName: Constants.Timer.Symbol.stretching)
           .resizable()
@@ -28,6 +30,8 @@ struct TimerTagLabelView: View {
   }
 }
 
+/// **타이머 정지(일시정지) 버튼**
+/// - 버튼을 누르면 타이머를 멈춤
 struct PauseButtonView: View {
   @EnvironmentObject var pomoVM: PomoViewModel
   
@@ -43,8 +47,3 @@ struct PauseButtonView: View {
     }
   }
 }
-
-//#Preview{
-//  TimerTagLabelView()
-//    .environmentObject(PomoViewModel(pomoTodoUseCase: DIContainer().pomoTodoUseCase))
-//}
