+++ PomoTodo/PomoTodo/Sources/Views/Pomo/ViewModel/PomoViewModel.swift
@@ -7,6 +7,8 @@
 
 import SwiftUI
 
+/// **포모도로 타이머의 ViewModel**
+/// - 타이머 설정, 상태 변경, 데이터 저장 및 UI 업데이트 관리
 final class PomoViewModel: ObservableObject {
   // MARK: - UI 변수
   /// 태그 관련
@@ -31,9 +33,10 @@ final class PomoViewModel: ObservableObject {
     }
   }
   
+  /// 현재 선택된 컬러 세트
   var selectedColorSet: TimerColorSet = Constants.Timer.indigoSet
   
-  /// 현재 타이머 상태 관련
+  /// 타이머 진행 상태 관련 변수
   @Published var curTomato = 1
   @Published var totalTomato = 2
   @Published var currentPhase: TimerPhase = .focus {
@@ -45,10 +48,12 @@ final class PomoViewModel: ObservableObject {
   @Published var isTimerRunning = false
   @Published var progress: CGFloat = 1.0
   
+  /// 현재 진행 중인 타이머의 총 시간
   @Published var totalTime: Int = 10
   @Published var remainingTime: Int = 10
   
   // MARK: - Action Case 정의
+  /// **포모도로 타이머에서 발생할 수 있는 이벤트**
   enum pomoAction {
     case onAppear
     case startTimer
@@ -56,14 +61,17 @@ final class PomoViewModel: ObservableObject {
     case forwardNextTimer
   }
   
-  // MARK: - Private 변수: 데이터 관련
-  private var pomoTodoUseCase: PomoTodoUseCase
+  // MARK: - internal 변수: 데이터 관련
+  internal var pomoTodoUseCase: PomoTodoUseCase
 
-  private var timer: Timer?
-  private var accumulatedFocusTime: Int = 0
-  private var accumulatedTotalTime: Int = 0
+  internal var timer: Timer?
+  internal var accumulatedFocusTime: Int = 0
+  internal var accumulatedTotalTime: Int = 0
   
   // MARK: - Initialization
+  
+  /// **포모도로 타이머 뷰모델 초기화**
+  /// - `pomoTodoUseCase`: 포모도로 데이터 관리 유즈케이스
   init (pomoTodoUseCase: PomoTodoUseCase) {
     self.pomoTodoUseCase = pomoTodoUseCase
     self.options = pomoTodoUseCase.getAppConfig().tags
@@ -71,6 +79,9 @@ final class PomoViewModel: ObservableObject {
   }
   
   // MARK: - Actions
+  
+  /// **이벤트 핸들러**
+  /// - `action`: 수행할 액션 (`pomoAction` 타입)
   func send(_ action: pomoAction) {
     switch action {
     case .onAppear:
@@ -85,6 +96,8 @@ final class PomoViewModel: ObservableObject {
   }
   
   // MARK: - Action Funcs
+  
+  /// **UI 초기 설정**
   func setUI() {
     self.options = pomoTodoUseCase.getAppConfig().tags
     self.timers = pomoTodoUseCase.getAppConfig().pomoTimers
@@ -93,6 +106,7 @@ final class PomoViewModel: ObservableObject {
     changeColorSet()
   }
   
+  /// **타이머 시작**
   func startTimer() {
     isTimerRunning = true
     remainingTime = totalTime
@@ -114,6 +128,7 @@ final class PomoViewModel: ObservableObject {
     }
   }
   
+  /// **타이머 중지**
   func stopTimer() {
     timer?.invalidate()
     saveFocusTime() // 시간 기록(토마토 완성 여부와 무관)
@@ -121,6 +136,7 @@ final class PomoViewModel: ObservableObject {
     progress = 1.0
   }
   
+  /// **다음 타이머 세션으로 이동**
   func forwardNextTimer() {
     if currentPhase == .focus {
       if curTomato < totalTomato {
@@ -147,52 +163,3 @@ final class PomoViewModel: ObservableObject {
   
 }
 
-extension PomoViewModel {
-  private func updateTotalTime() {
-    let selectedTimer = timers[currentPage]
-    
-    switch currentPhase {
-    case .focus:
-      totalTime = selectedTimer.focusTimeUnit.asInt
-    case .shortBreak:
-      totalTime = selectedTimer.shortBreakUnit.asInt
-    case .longBreak:
-      totalTime = selectedTimer.longBreakUnit.asInt
-    }
-    
-    remainingTime = totalTime
-    totalTomato = selectedTimer.tomatoPerCycle
-    progress = 1.0
-  }
-  
-  private func triggerHapticFeedback() {
-      let generator = UINotificationFeedbackGenerator()
-      generator.notificationOccurred(.success)
-  }
-  
-  private func resetCycle() {
-    curTomato = 1 // 암튼 무조건 세션 초기화
-  }
-  
-  private func changeColorSet() {
-    selectedColorSet = Constants.Timer.colorSets.first { $0.id == options[selectionTag].colorId } ?? Constants.Timer.indigoSet
-  }
-  
-  // MARK: - Data Funcs
-  // 집중 시간이랑 전체 시간(집중 + 휴식) 저장
-  private func saveFocusTime() {
-    if currentPhase == .focus {
-      accumulatedFocusTime += (totalTime - remainingTime)
-      pomoTodoUseCase.addTagTimeRecords(todayPomoDay: pomoTodoUseCase.getTodayPomoDay(), tagTimeRecord: TagTimeRecord(tagId: options[selectionTag].id, focusTime: accumulatedFocusTime.asDouble))
-      accumulatedFocusTime = 0
-    } else {
-      accumulatedTotalTime += (totalTime - remainingTime)
-    }
-  }
-  
-  private func saveFocusTomato() {
-    if currentPhase == .focus {
-      pomoTodoUseCase.updateTomatoAndCycle(todayPomoDay: pomoTodoUseCase.getTodayPomoDay(), tomatoCnt: 1, cycleCnt: 1 / totalTomato.asDouble)
-    }
-  }
-}
